reductions and subspaces
